split userDataUseCase
