+++ app/src/main/java/com/space/quizapp/domain/usecase/auth/GetCurrentUserIdUseCase.kt
@@ -1,0 +1,8 @@
+package com.space.quizapp.domain.usecase.auth
+
+import com.space.quizapp.domain.repository.AuthenticationRepository
+import javax.inject.Inject
+
+class GetCurrentUserIdUseCase @Inject constructor(private val authenticationRepository: AuthenticationRepository) {
+    fun invoke() = authenticationRepository.getCurrentUserId()
+}
