+++ app/src/main/java/com/space/quizapp/presentation/point/vm/PointsViewModel.kt
@@ -4,10 +4,10 @@ import androidx.lifecycle.viewModelScope
 import com.space.quizapp.common.extensions.toResult
 import com.space.quizapp.common.mapper.toUIModel
 import com.space.quizapp.common.resource.Result
-import com.space.quizapp.domain.usecase.auth.AuthenticationUseCase
-import com.space.quizapp.domain.usecase.auth.GetCurrentUseIdUseCase
+import com.space.quizapp.domain.usecase.auth.GetCurrentUserIdUseCase
 import com.space.quizapp.domain.usecase.auth.LogOutUseCase
-import com.space.quizapp.domain.usecase.user.UserDataUseCse
+import com.space.quizapp.domain.usecase.user.GetUserDataUseCse
+import com.space.quizapp.domain.usecase.user.GetUserPointsUseCse
 import com.space.quizapp.presentation.base.vm.BaseViewModel
 import com.space.quizapp.presentation.home.ui.HomeFragmentDirections
 import com.space.quizapp.presentation.model.PointUIModel
@@ -21,9 +21,9 @@ import javax.inject.Inject
 
 @HiltViewModel
 class PointsViewModel @Inject constructor(
-    private val getCurrentUseIdUseCase: GetCurrentUseIdUseCase,
+    private val getCurrentUseIdUseCase: GetCurrentUserIdUseCase,
     private val logOutUseCase: LogOutUseCase,
-    private val userDataUseCse: UserDataUseCse,
+    private val getUserPointsUseCse: GetUserPointsUseCse
 ) : BaseViewModel() {
 
     private val currentUserId = getCurrentUseIdUseCase.invoke()
@@ -46,7 +46,7 @@ class PointsViewModel @Inject constructor(
 
     private fun getPoints() {
         viewModelScope.launch {
-            userDataUseCse.getUserPoints(currentUserId).map {
+            getUserPointsUseCse.invoke(currentUserId).map {
                 it.map { point -> point.toUIModel() }
             }.toResult().collectLatest {
                 _points.tryEmit(it)
